requirements.txt
Django
pytz
psycopg2
djangorestframework
django-filter